+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Drag images from a chat and drop it to another chat or even to another app (use a second finger for navigating!) (#2509)
 - Paste .webp images via drag'n'drop (#2507)
 - Add "Copy Image" to messages' conext menu
+- Long-tap a proxy to share or delete (#2521)
 - "Message Info" is moved to messages' context menu at "More Options / ..." (#2510)
 - Allow cancelling profile edits, force a name when editing profile (#2506)
 - Fix scrolling issue when cancelling a screen (#2504)